+++ AccountManagement/PhotoData.h
@@ -12,7 +12,7 @@
 @interface PhotoData : NSObject
 
 @property (nonatomic) int pid;
-@property (nonatomic) long timeInterval;
+@property (nonatomic) long long timeInterval;
 @property (nonatomic) int abid;
 @property (nonatomic) int sync;
